+++ packages/expo-router/CHANGELOG.md
@@ -60,6 +60,7 @@
 - Upgrade rnscreens to nigthly version and remove RNSDismissibleModalProtocol patch ([#38522](https://github.com/expo/expo/pull/38522) by [@Ubax](https://github.com/Ubax))
 - Move not-found and site map to root stack navigator ([#38417](https://github.com/expo/expo/pull/38417) by [@Ubax](https://github.com/Ubax))
 - Hide standalone Modal export ([#38648](https://github.com/expo/expo/pull/38648) by [@Ubax](https://github.com/Ubax))
+- Refactor native tabs for docs export([#38684](https://github.com/expo/expo/pull/38684) by [@Ubax](https://github.com/Ubax))
 
 ## 5.1.4 - 2025-07-18